// client/src/lib/movies.ts
export type MoviesType = {
    map(arg0: (movie: MoviesType) => import("react/jsx-runtime").JSX.Element): import("react").ReactNode | Iterable<import("react").ReactNode>;
    idmovie: number;
    title_en: string;
    title_pt: string;
    release_date?: string;
    fk_producer01?: number; 
    fk_producer02?: number; 
    fk_producer03?: number; 
    fk_producer04?: number; 
    fk_producer05?: number; 
    fk_director01?: number; 
    fk_director02?: number; 
    fk_director03?: number; 
    fk_director04?: number; 
    fk_director05?: number; 
    fk_writer01?: number; 
    fk_writer02?: number; 
    fk_writer03?: number; 
    fk_writer04?: number; 
    fk_writer05?: number; 
    fk_actor01?: number;
    fk_actor02?: number;
    fk_actor03?: number;
    fk_actor04?: number;
    fk_actor05?: number;
    fk_actor06?: number;
    fk_actor07?: number;
    fk_actor08?: number;
    fk_actor09?: number;
    fk_actor10?: number;
    fk_actor11?: number;
    fk_actor12?: number;
    fk_actor13?: number;
    fk_actor14?: number;
    fk_actor15?: number;
    fk_actor16?: number;
    fk_actor17?: number;
    fk_actor18?: number;
    fk_actor19?: number;
    fk_actor20?: number;
    fk_genre_en01?: number; 
    fk_genre_en02?: number; 
    fk_genre_en03?: number; 
    fk_genre_en04?: number; 
    fk_genre_en05?: number; 
    fk_genre_pt01?: number; 
    fk_genre_pt02?: number; 
    fk_genre_pt03?: number; 
    fk_genre_pt04?: number; 
    fk_genre_pt05?: number; 
    fk_rated_pg_en?: number; 
    fk_rated_pg_pt?: number;
    runtime?: string;
    budget?: number;
    box_office?: number;
    about_movie_en?: string;
    about_movie_pt?: string;
    cover_movie_url: string;
    trailer_movie_url?: string;
    image_movie_url01?: string; 
    image_movie_url02?: string; 
    image_movie_url03?: string; 
    image_movie_url04?: string; 
    image_movie_url05?: string; 
    image_movie_url06?: string; 
    image_movie_url07?: string; 
    image_movie_url08?: string; 
    image_movie_url09?: string; 
    image_movie_url10?: string;
    my_rating?: number;
    my_review_en?: string;
    my_review_pt?: string;
}